+++ contracts/subscriptions/AragonAppFeesCashier.sol
@@ -4,10 +4,10 @@ import "../court/controller/Controller.sol";
 import "../court/controller/Controlled.sol";
 import "../lib/os/EtherTokenConstant.sol";
 
-import "./ITransactionFeesOracle.sol";
+import "./IAragonAppFeesCashier.sol";
 
 
-contract TransactionFeesOracle is Controlled, ITransactionFeesOracle, EtherTokenConstant {
+contract AragonAppFeesCashier is Controlled, IAragonAppFeesCashier, EtherTokenConstant {
     string private constant ERROR_APP_ID_NOT_SET = "SFO_APP_NOT_SET";
     string private constant ERROR_WRONG_TOKEN = "SFO_WRONG_TOKEN";
     string private constant ERROR_COURT_HAS_NOT_STARTED = "SFO_COURT_HAS_NOT_STARTED";
@@ -36,8 +36,8 @@ contract TransactionFeesOracle is Controlled, ITransactionFeesOracle, EtherToken
     * @param _token Token for the fee
     * @param _amount Amount of fee tokens
     */
-    function setTransactionFee(bytes32 _appId, ERC20 _token, uint256 _amount) external onlyConfigGovernor {
-        _setTransactionFee(_appId, _token, _amount);
+    function setAppFee(bytes32 _appId, ERC20 _token, uint256 _amount) external onlyConfigGovernor {
+        _setAppFee(_appId, _token, _amount);
     }
 
     /**
@@ -46,12 +46,12 @@ contract TransactionFeesOracle is Controlled, ITransactionFeesOracle, EtherToken
     * @param _tokens Token for the fees for each app
     * @param _amounts Amount of fee tokens for each app
     */
-    function setTransactionFees(bytes32[] calldata _appIds, ERC20[] calldata _tokens, uint256[] calldata _amounts) external onlyConfigGovernor {
+    function setAppFees(bytes32[] calldata _appIds, ERC20[] calldata _tokens, uint256[] calldata _amounts) external onlyConfigGovernor {
         require(_appIds.length == _tokens.length, ERROR_WRONG_TOKENS_LENGTH);
         require(_appIds.length == _amounts.length, ERROR_WRONG_AMOUNTS_LENGTH);
 
         for (uint256 i = 0; i < _appIds.length; i++) {
-            _setTransactionFee(_appIds[i], _tokens[i], _amounts[i]);
+            _setAppFee(_appIds[i], _tokens[i], _amounts[i]);
         }
     }
 
@@ -59,23 +59,23 @@ contract TransactionFeesOracle is Controlled, ITransactionFeesOracle, EtherToken
     * @notice Unset fees for app with id `_appId`
     * @param _appId Id of the app
     */
-    function unsetTransactionFee(bytes32 _appId) external onlyConfigGovernor {
-        _unsetTransactionFee(_appId);
+    function unsetAppFee(bytes32 _appId) external onlyConfigGovernor {
+        _unsetAppFee(_appId);
     }
 
     /**
     * @notice Unset fees for apps with ids `_appIds`
     * @param _appIds Ids of the apps
     */
-    function unsetTransactionFees(bytes32[] calldata _appIds) external onlyConfigGovernor {
+    function unsetAppFees(bytes32[] calldata _appIds) external onlyConfigGovernor {
         for (uint256 i = 0; i < _appIds.length; i++) {
-            _unsetTransactionFee(_appIds[i]);
+            _unsetAppFee(_appIds[i]);
         }
     }
 
     // TODO: To be integrated with CourtSubscriptions with the new trusted model
-    function payTransactionFees(bytes32 _appId, uint256 _actionId) external {
-        emit TransactionFeePaid(msg.sender, _appId, _actionId);
+    function payAppFees(bytes32 _appId, uint256 _actionId) external {
+        emit AppFeePaid(msg.sender, _appId, _actionId);
     }
 
     /**
@@ -84,7 +84,7 @@ contract TransactionFeesOracle is Controlled, ITransactionFeesOracle, EtherToken
     * @return Token for the fees
     * @return Amount of fee tokens
     */
-    function getTransactionFee(bytes32 _appId) external view returns (ERC20 token, uint256 amount) {
+    function getAppFee(bytes32 _appId) external view returns (ERC20 token, uint256 amount) {
         AppFee storage appFee = appFees[_appId];
 
         require(appFee.set, ERROR_APP_ID_NOT_SET);
@@ -92,7 +92,7 @@ contract TransactionFeesOracle is Controlled, ITransactionFeesOracle, EtherToken
         return (appFee.token, appFee.amount);
     }
 
-    function _setTransactionFee(bytes32 _appId, ERC20 _token, uint256 _amount) internal {
+    function _setAppFee(bytes32 _appId, ERC20 _token, uint256 _amount) internal {
         require(address(_token) == ETH || isContract(address(_token)), ERROR_WRONG_TOKEN);
 
         AppFee storage appFee = appFees[_appId];
@@ -101,14 +101,14 @@ contract TransactionFeesOracle is Controlled, ITransactionFeesOracle, EtherToken
         appFee.token = _token;
         appFee.amount = _amount;
 
-        emit TransactionFeeSet(_appId, _token, _amount);
+        emit AppFeeSet(_appId, _token, _amount);
     }
 
-    function _unsetTransactionFee(bytes32 _appId) internal {
+    function _unsetAppFee(bytes32 _appId) internal {
         require(appFees[_appId].set, ERROR_APP_ID_NOT_SET);
 
         delete appFees[_appId];
 
-        emit TransactionFeeUnset(_appId);
+        emit AppFeeUnset(_appId);
     }
 }
